pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "research-proj-template"
version = "0.1.0"
description = "Reusable research-to-infra project template."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "your_email@example.com" },
]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/research_proj_template"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
